More Refactor of props and properties
Improving implementation for less required editing on different configurations.
