+++ config_robot/suitcase/501robot.props
@@ -7,6 +7,7 @@
 #########################################
 
 Robot.name = Suitcase-Bot
+Robot.implementation = Suitcase
