+++ src/main/java/frc/robot/properties/package-info.java
@@ -6,19 +6,18 @@
 /* of this project.                                                      */
 /*-----------------------------------------------------------------------*/
 
-// WARNING: This file is automatically generated in build.gradle
-
-package frc.robot.config;
-
 /**
- * This class provides a versioning string that can be used to determine the
- * build pedigree of a robot as loaded on the roboRIO.
+ * This package contains the implementation of our preferences.
  * <p>
- * Unfortunately, Git doesn't provide a way to automate this the way other CM
- * tools do with keyword expansions.
+ * <code>Properties</code> are values used by the robot code to configure
+ * and provide change capability without modifying and re-deploying code.
+ * However, unlike <code>Preferences</code> they can not be manipulated
+ * except by editing the contents of the file on the robot itself, and
+ * restarting the robot code.
+ * </p>
+ *
+ * @since 2022.0.0
+ * @author first.stu
+ * @version 2022.0.0
  **/
-class BuildVersionInfo {
-    public static final String programmer = "Stu Lewin";
-    public static final String commitSHA = "32b9464";
-    public static final String timeStamp = "20220206-161341";
-}
+package frc.robot.properties;
