+++ src/main/java/frc/robot/sensors/SensorFactory.java
@@ -15,8 +15,6 @@ import java.util.List;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 import frc.robot.sensors.gyro.GyroFactory;
-import frc.robot.sensors.vision.VisionFactory;
-import frc.robot.sensors.turretlocation.TurretLocationFactory;
 import frc.robot.telemetry.TelemetryManager;
 import frc.robot.telemetry.TelemetryNames;
 import frc.robot.utils.PKStatus;
@@ -44,22 +42,6 @@ public class SensorFactory {
         {
             GyroFactory.constructInstance();
             ISensor s = GyroFactory.getInstance();
-            tlmMgr.addProvider(s);
-            sensors.add(s);
-        }
-
-        SmartDashboard.putNumber(TelemetryNames.TurretLocation.status, PKStatus.unknown.tlmValue);
-        {
-            TurretLocationFactory.constructInstance();
-            ISensor s = TurretLocationFactory.getInstance();
-            tlmMgr.addProvider(s);
-            sensors.add(s);
-        }
-
-        SmartDashboard.putNumber(TelemetryNames.Vision.status, PKStatus.unknown.tlmValue);
-        {
-            VisionFactory.constructInstance();
-            ISensor s = VisionFactory.getInstance();
             tlmMgr.addProvider(s);
             sensors.add(s);
         }
